+++ internal/service/cloudcontrol/resource.go
@@ -11,6 +11,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/cloudcontrolapi"
 	cfschema "github.com/hashicorp/aws-cloudformation-resource-schema-sdk-go"
+	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -121,7 +122,7 @@ func resourceResourceRead(ctx context.Context, d *schema.ResourceData, meta inte
 	conn := meta.(*conns.AWSClient).CloudControlConn
 
 	typeName := d.Get("type_name").(string)
-	resourceDescription, err := FindResourceByID(ctx, conn,
+	resourceDescription, err := FindResource(ctx, conn,
 		d.Id(),
 		typeName,
 		d.Get("type_version_id").(string),
@@ -306,6 +307,109 @@ func resourceResourceCustomizeDiffSchemaDiff(ctx context.Context, diff *schema.R
 	}
 
 	return nil
+}
+
+func FindResource(ctx context.Context, conn *cloudcontrolapi.CloudControlApi, resourceID, typeName, typeVersionID, roleARN string) (*cloudcontrolapi.ResourceDescription, error) {
+	input := &cloudcontrolapi.GetResourceInput{
+		Identifier: aws.String(resourceID),
+		TypeName:   aws.String(typeName),
+	}
+	if roleARN != "" {
+		input.RoleArn = aws.String(roleARN)
+	}
+	if typeVersionID != "" {
+		input.TypeVersionId = aws.String(typeVersionID)
+	}
+
+	output, err := conn.GetResourceWithContext(ctx, input)
+
+	if tfawserr.ErrCodeEquals(err, cloudcontrolapi.ErrCodeResourceNotFoundException) {
+		return nil, &resource.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	// Some CloudFormation Resources do not correctly re-map "not found" errors, instead returning a HandlerFailureException.
+	// These should be reported and fixed upstream over time, but for now work around the issue.
+	if tfawserr.ErrMessageContains(err, cloudcontrolapi.ErrCodeHandlerFailureException, "not found") {
+		return nil, &resource.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil || output.ResourceDescription == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output.ResourceDescription, nil
+}
+
+func findProgressEventByRequestToken(ctx context.Context, conn *cloudcontrolapi.CloudControlApi, requestToken string) (*cloudcontrolapi.ProgressEvent, error) {
+	input := &cloudcontrolapi.GetResourceRequestStatusInput{
+		RequestToken: aws.String(requestToken),
+	}
+
+	output, err := conn.GetResourceRequestStatusWithContext(ctx, input)
+
+	if tfawserr.ErrCodeEquals(err, cloudcontrolapi.ErrCodeRequestTokenNotFoundException) {
+		return nil, &resource.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil || output.ProgressEvent == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output.ProgressEvent, nil
+}
+
+func statusProgressEventOperation(ctx context.Context, conn *cloudcontrolapi.CloudControlApi, requestToken string) resource.StateRefreshFunc {
+	return func() (interface{}, string, error) {
+		output, err := findProgressEventByRequestToken(ctx, conn, requestToken)
+
+		if tfresource.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		return output, aws.StringValue(output.OperationStatus), nil
+	}
+}
+
+func waitProgressEventOperationStatusSuccess(ctx context.Context, conn *cloudcontrolapi.CloudControlApi, requestToken string, timeout time.Duration) (*cloudcontrolapi.ProgressEvent, error) {
+	stateConf := &resource.StateChangeConf{
+		Pending: []string{cloudcontrolapi.OperationStatusInProgress, cloudcontrolapi.OperationStatusPending},
+		Target:  []string{cloudcontrolapi.OperationStatusSuccess},
+		Refresh: statusProgressEventOperation(ctx, conn, requestToken),
+		Timeout: timeout,
+	}
+
+	outputRaw, err := stateConf.WaitForStateContext(ctx)
+
+	if output, ok := outputRaw.(*cloudcontrolapi.ProgressEvent); ok {
+		if operationStatus := aws.StringValue(output.OperationStatus); operationStatus == cloudcontrolapi.OperationStatusFailed {
+			tfresource.SetLastError(err, fmt.Errorf("%s: %s", aws.StringValue(output.ErrorCode), aws.StringValue(output.StatusMessage)))
+		}
+
+		return output, err
+	}
+
+	return nil, err
 }
 
 // patchDocument returns a JSON Patch document describing the difference between `old` and `new`.
